Fix version number logic
Previously this was using the build in Kubernetes style version
handling.  That meant that the version number was always wrong and that
it claimed to be Kubernetes when it printed the version number.  In
addition, it's hard to change this since the program name is hard coded
in the logic to print out the version number.

To fix those issues and allow logging the version number when start,
this patch changes the version handling logic to work internally.

This also patches the Makefile to properly escape the *.go pattern in
the verify target so that it isn't expanded early.